first approach of new Load Safe flow
