--- src/routes/new-load/constants/load-field-names.tsx
@@ -1,3 +1,0 @@
-export const FIELD_LOAD_CUSTOM_SAFE_NAME = 'customSafeName'
-export const FIELD_LOAD_SUGGESTED_SAFE_NAME = 'suggestedSafeName'
-export const FIELD_LOAD_SAFE_ADDRESS = 'safeAddress'
