+++ src/routes/new-load/steps/LoadSafeOwnersStep.tsx
@@ -2,6 +2,8 @@ import React, { ReactElement } from 'react'
 import { useForm } from 'react-final-form'
 import { useStepper } from 'src/components/NewStepper/stepperContext'
 
+export const loadSafeOwnersStepLabel = 'Owners'
+
 function LoadSafeOwnersStep(): ReactElement {
   const loadSafeForm = useForm()
   const loadSafeStepper = useStepper()
@@ -14,3 +16,9 @@ function LoadSafeOwnersStep(): ReactElement {
 }
 
 export default LoadSafeOwnersStep
+
+export const loadSafeOwnersStepValidation = (values) => {
+  const errors = {}
+  console.log('Validations for Owners', values)
+  return errors
+}
