+++ src/routes/new-load/steps/ReviewLoadStep.tsx
@@ -1,7 +1,19 @@
 import React, { ReactElement } from 'react'
 
+export const reviewLoadStepLabel = 'Review'
+
 function ReviewLoadStep(): ReactElement {
-  return <div>Last step: ReviewLoadStep</div>
+  return (
+    <div>
+      <div>LAST step: ReviewLoadStep</div>
+    </div>
+  )
 }
 
 export default ReviewLoadStep
+
+export const reviewLoadStepValidations = (values) => {
+  const errors = {}
+  console.log('Validations for Review', values)
+  return errors
+}
